+++ app/Models/Blueprint.php
@@ -6,5 +6,12 @@ use Illuminate\Database\Eloquent\Model;
 
 class Blueprint extends Model
 {
-    //
+    public function user() : \Illuminate\Database\Eloquent\Relations\BelongsTo
+    {
+        return $this->belongsTo(User::class);
+    }
+    public function category() : \Illuminate\Database\Eloquent\Relations\BelongsTo
+    {
+        return $this->belongsTo(Category::class);
+    }
 }