+++ app/Models/Category.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Category extends Model
+{
+    public function blueprints() : \Illuminate\Database\Eloquent\Relations\HasMany
+    {
+        return $this->hasMany(Blueprint::class);
+    }
+}