+++ changelog.md
@@ -15,12 +15,37 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 [//]: # (### Removed)
 [//]: # (- item)
 
+## 15/10/2025
+### Added
+- Create view for blueprints
+- Edit view for blueprints
+- Store functionality for blueprints
+- Delete functionality for blueprints
+- Update functionality for blueprints management
+- Validation in BlueprintController
+- Category table, model, migration and controller
+- Relationship between blueprints and categories
+- Textarea component
+
+### Changed
+- Blueprints table schema (timestamps, categories added)
+- Show view for blueprints
+
 ## 14/10/2025
 ### Added
-add Blueprint model, migrations for blueprints, tags, and likes; update views and routing
+- Footer 
+- Blueprint model
+- Migrations for blueprints tags, and likes
 
 ## 13/10/2025
 ### Added
+- Navigation bar
+
+### Changed
+- rename about-us view to about
+- update routing
+- update views
+- updated name in .env.example
 
 ## 08/10/2025
 ### Added